refactoring: moved old code to Controls folder
