+++ view/desktop/src/components/HeadBar/Controls/icons.tsx
