+++ view/desktop/src/components/index.ts
@@ -1,9 +1,6 @@
 export * from "./AppLayout";
 export * from "./Menu";
 export * from "./Button";
-export * from "./window-controls/WindowControls";
-export * from "./WindowTitlebar";
-export * from "./TitleBar";
 export * from "./LanguageSelector";
 export * from "./ThemeSelector";
 export * from "./Sidebar";
@@ -12,3 +9,5 @@ export * from "./SidebarHeader";
 export * from "./SidebarGeneral";
 export * from "./SidebarLinks";
 export * from "./LaunchPad";
+export * from "./Resizable";
+export * from "./HeadBar";
