--- view/desktop/src/components/window-controls/contexts/index.ts
@@ -1,2 +1,0 @@
-export { TauriAppWindowProvider } from "./plugin-window";
-export { TauriAppWindowContext } from "./plugin-window";
