--- view/desktop/src/components/window-controls/controls/index.ts
@@ -1,3 +1,0 @@
-export { Windows } from "./windows";
-export { MacOS } from "./macos";
-export { Gnome } from "./linux";
