--- view/desktop/src/components/window-controls/controls/linux/index.ts
@@ -1,1 +1,0 @@
-export { Gnome } from "./gnome";
