+++ view/desktop/src/main.tsx
@@ -7,10 +7,10 @@ import { store } from "./store";
 import { type } from "@tauri-apps/plugin-os";
 
 if (type() !== "windows") {
+  document.querySelectorAll("html, body").forEach((el) => {
+    el.classList.add("rounded-2xl");
+  });
 }
-document.querySelectorAll("html, body").forEach((el) => {
-  el.classList.add("rounded-2xl");
-});
 
 createRoot(document.getElementById("root") as HTMLElement).render(
   <StrictMode>
